Support preloading through `maxPreloadedNodes` config option
Dagon will now look at the links between nodes as defined by the game and construct a control flow graph from them.
After switching to a node, Dagon will do a BFS traversal of the control flow graph starting from the current node. Nodes will be preloaded in their BFS order.
If the preloader runs out of nodes to preload, but it's still not at the `maxPreloadedNodes` threshold, it will begin preloading nodes of the current room (in arbitrary order).
When the threshold is exceeded, nodes will be unloaded in the order they were preloaded. i.e. the first one in is the first one out.
`maxPreloadedNodes` defaults to 3.