Preloading take 2
Like before, Dagon will build a graph out of links between nodes and use this to do preloading.
The difference is that now preloading is not limited by a number of nodes, but by a distance from the current node.
Another difference is when textures are unloaded. After a switch, any spot/node-texture that is not used within preloading distance will be unloaded.
`config.preloadDistance` replaces `config.maxPreloadedNodes` with this commit.
`config.preloadDistance` defaults to 1. Set to 0 to disable preloading.